Fix Issue84 + remove vendored files from git + move servertest to app dir
